Rename package common to model
