+++ internal/verification/verifier_test.go
@@ -19,7 +19,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/project-oak/transparent-release/internal/common"
+	"github.com/project-oak/transparent-release/internal/model"
 	slsav02 "github.com/project-oak/transparent-release/pkg/intoto/slsa_provenance/v0.2"
 )
 
@@ -30,7 +30,7 @@ const (
 
 func TestVerify_HasNoValues(t *testing.T) {
 	// There are no optional fields set apart from the binary digest and the build type.
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName)
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName)
 
 	want := ReferenceValues{
 		// We ask for all the optional values in the reference values.
@@ -53,7 +53,7 @@ func TestVerify_HasNoValues(t *testing.T) {
 }
 
 func TestVerify_HasBuildCmd_HasAndNeedsBuildCmd(t *testing.T) {
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuildCmd([]string{"build cmd"}))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuildCmd([]string{"build cmd"}))
 
 	want := ReferenceValues{
 		WantBuildCmds: true,
@@ -71,7 +71,7 @@ func TestVerify_HasBuildCmd_HasAndNeedsBuildCmd(t *testing.T) {
 
 func TestVerify_NeedsButCannotHaveNoBuildCmd(t *testing.T) {
 	// No buildCmd is set in the provenance.
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName)
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName)
 
 	want := ReferenceValues{
 		WantBuildCmds: true,
@@ -89,7 +89,7 @@ func TestVerify_NeedsButCannotHaveNoBuildCmd(t *testing.T) {
 
 func TestVerify_NeedsButHasNoBuildCmd(t *testing.T) {
 	// The build command is empty.
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuildCmd([]string{}))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuildCmd([]string{}))
 	// And the reference values ask for a build cmd.
 	want := ReferenceValues{
 		WantBuildCmds: true,
@@ -108,7 +108,7 @@ func TestVerify_NeedsButHasNoBuildCmd(t *testing.T) {
 
 func TestVerify_HasNoBuildCmdButNotNeeded(t *testing.T) {
 	// The build command is empty.
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuildCmd([]string{}))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuildCmd([]string{}))
 	// But the reference values do not ask for a build cmd.
 	want := ReferenceValues{
 		WantBuildCmds: false,
@@ -127,7 +127,7 @@ func TestVerify_HasNoBuildCmdButNotNeeded(t *testing.T) {
 
 func TestVerify_HasAndNeedsBuilderImageDigest(t *testing.T) {
 	builderDigest := "9e2ba52487d945504d250de186cb4fe2e3ba023ed2921dd6ac8b97ed43e76af9"
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuilderImageSHA256Digest(builderDigest))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuilderImageSHA256Digest(builderDigest))
 	want := ReferenceValues{
 		BuilderImageSHA256Digests: []string{"some_other_digest", builderDigest},
 	}
@@ -144,7 +144,7 @@ func TestVerify_HasAndNeedsBuilderImageDigest(t *testing.T) {
 
 func TestVerify_NeedsButBuilderImageDigestNotFound(t *testing.T) {
 	builderDigest := "9e2ba52487d945504d250de186cb4fe2e3ba023ed2921dd6ac8b97ed43e76af9"
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuilderImageSHA256Digest(builderDigest))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuilderImageSHA256Digest(builderDigest))
 	want := ReferenceValues{
 		BuilderImageSHA256Digests: []string{"some_other_digest", "and_some_other"},
 	}
@@ -163,7 +163,7 @@ func TestVerify_NeedsButBuilderImageDigestNotFound(t *testing.T) {
 }
 
 func TestVerify_NeedsButHasEmptyBuilderImageDigest(t *testing.T) {
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuilderImageSHA256Digest(""))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuilderImageSHA256Digest(""))
 	want := ReferenceValues{
 		BuilderImageSHA256Digests: []string{"some_digest"},
 	}
@@ -183,7 +183,7 @@ func TestVerify_NeedsButHasEmptyBuilderImageDigest(t *testing.T) {
 
 func TestVerify_HasEmptyBuilderImageDigestButNotNeeded(t *testing.T) {
 	builderImageSHA256Digest := ""
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithBuilderImageSHA256Digest(builderImageSHA256Digest))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithBuilderImageSHA256Digest(builderImageSHA256Digest))
 	want := ReferenceValues{
 		// We do not check for the builder image digest.
 	}
@@ -199,8 +199,8 @@ func TestVerify_HasEmptyBuilderImageDigestButNotNeeded(t *testing.T) {
 }
 
 func TestVerify_HasFoundRepoURI(t *testing.T) {
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName,
-		common.WithRepoURIs([]string{
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName,
+		model.WithRepoURIs([]string{
 			"git+https://github.com/project-oak/transparent-release@refs/heads/main",
 			"https://github.com/project-oak/transparent-release",
 		}))
@@ -221,9 +221,9 @@ func TestVerify_HasFoundRepoURI(t *testing.T) {
 
 func TestVerify_HasWrongRepoURI(t *testing.T) {
 	wrongURI := "git+https://github.com/project-oak/oak@refs/heads/main"
-	got := common.NewProvenanceIR(binarySHA256Digest,
+	got := model.NewProvenanceIR(binarySHA256Digest,
 		slsav02.GenericSLSABuildType, binaryName,
-		common.WithRepoURIs([]string{
+		model.WithRepoURIs([]string{
 			wrongURI,
 			"https://github.com/project-oak/transparent-release",
 		}))
@@ -247,8 +247,8 @@ func TestVerify_HasWrongRepoURI(t *testing.T) {
 
 func TestVerify_HasNoRepoURIs(t *testing.T) {
 	// We have no repo URIs in the provenance.
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName,
-		common.WithRepoURIs([]string{}))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName,
+		model.WithRepoURIs([]string{}))
 	want := ReferenceValues{
 		RepoURI: "github.com/project-oak/transparent-release",
 	}
@@ -266,7 +266,7 @@ func TestVerify_HasNoRepoURIs(t *testing.T) {
 
 func TestVerify_HasAndNeedsTrustedBuilder(t *testing.T) {
 	trustedBuilder := "https://github.com/slsa-framework/slsa-github-generator/.github/workflows/generator_generic_slsa3.yml@refs/tags/v1.2.0"
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithTrustedBuilder(trustedBuilder))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithTrustedBuilder(trustedBuilder))
 
 	want := ReferenceValues{
 		TrustedBuilders: []string{trustedBuilder},
@@ -284,7 +284,7 @@ func TestVerify_HasAndNeedsTrustedBuilder(t *testing.T) {
 
 func TestVerify_NeedsButTrustedBuilderNotFound(t *testing.T) {
 	trustedBuilder := "https://github.com/slsa-framework/slsa-github-generator/.github/workflows/generator_generic_slsa3.yml@refs/tags/v1.2.0"
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithTrustedBuilder(trustedBuilder))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithTrustedBuilder(trustedBuilder))
 
 	want := ReferenceValues{
 		TrustedBuilders: []string{"other_" + trustedBuilder, "another_" + trustedBuilder},
@@ -304,7 +304,7 @@ func TestVerify_NeedsButTrustedBuilderNotFound(t *testing.T) {
 }
 
 func TestVerify_NeedsButHasEmptyTrustedBuilder(t *testing.T) {
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithTrustedBuilder(""))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithTrustedBuilder(""))
 
 	want := ReferenceValues{
 		TrustedBuilders: []string{"other_trusted_builder", "another_trusted_builder"},
@@ -324,7 +324,7 @@ func TestVerify_NeedsButHasEmptyTrustedBuilder(t *testing.T) {
 }
 
 func TestVerify_HasEmptyTrustedBuilderButNotNeeded(t *testing.T) {
-	got := common.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, common.WithTrustedBuilder(""))
+	got := model.NewProvenanceIR(binarySHA256Digest, slsav02.GenericSLSABuildType, binaryName, model.WithTrustedBuilder(""))
 
 	want := ReferenceValues{
 		// We do not check the trusted builder.
